% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{assign_lmc_sills}
\alias{assign_lmc_sills}
\title{Assign the LMC sills back to the gstat object}
\usage{
assign_lmc_sills(g, sills)
}
\arguments{
\item{g}{gstat object}
}
\value{
sills produced from LMC algorithm
}
\description{
Assign the LMC sills back to the gstat object
}
