% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate weights for variogram for simulated annealing
Ported from Emery at al. 2009
Added log methods to soften later portion of curve}
\usage{
calculate_weights(weighting_method, gam)
}
\arguments{
\item{weighting_method}{0:constant weight, 1: ~npairs, 2:~/distance 3: ~npairs/distance, 4: log(~npairs), 5:~/log(distance) 6: ~log(npairs)/log(distance)}

\item{gam}{the variogram/correlogram}
}
\value{
vector of weights, equal to nrows of variogram
}
\description{
Calculate weights for variogram for simulated annealing
Ported from Emery at al. 2009
Added log methods to soften later portion of curve
}
