% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{check_pull_SubString}
\alias{check_pull_SubString}
\title{check_pull_SubString}
\usage{
check_pull_SubString(asciitxt, sub.str)
}
\arguments{
\item{asciitxt}{las ascii object}

\item{sub.str}{pattern to search for}
}
\value{

}
\description{
check for a pattern in the LAS-as-ascii
}
