% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{classify_log_intervals}
\alias{classify_log_intervals}
\title{Mappable function to subset logs into shale and carbonate based on tops
works on elev_masl (what the tops are picked in)}
\usage{
classify_log_intervals(this_uwi, comb_las_df, dv_carb_tops)
}
\arguments{
\item{this_uwi}{the uwi for subsetting and facies classification}

\item{comb_las_df}{the entire combined las df}

\item{dv_carb_tops}{the tops of the Duvernay (has start and end of carbonate)}
}
\value{
a dataframe of a single well, to be combined using purrr::map_df
}
\description{
Mappable function to subset logs into shale and carbonate based on tops
works on elev_masl (what the tops are picked in)
}
