% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{cova}
\alias{cova}
\title{Compute covariance for reduced distance h
Ported from Emery at al. 2009}
\usage{
cova(type, h, epsilon = 1e-12)
}
\arguments{
\item{type}{0:constant, 1:nugget, 2:spherical, 3:exponential, 4:gaussian, else: linear}

\item{h}{lag distance}

\item{epsilon}{absolute comparison accuracy}
}
\value{
covariance based on reduced lag distance h
}
\description{
Compute covariance for reduced distance h
Ported from Emery at al. 2009
}
