% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{fit_linear_model_of_coregionalization}
\alias{fit_linear_model_of_coregionalization}
\title{Fit Linear Model of Coregionalization
Ported from Emergy et al. 2009}
\usage{
fit_linear_model_of_coregionalization(
  azimuth,
  dip,
  nlag,
  tail,
  head,
  gam,
  model,
  vartype,
  maxiterations = 5000,
  max_time_s = 100,
  p_acc_0 = 0.9,
  p_acc_2k = 0.1,
  gen_vec_corr = 0.9,
  eps_thresh = 1e-12,
  weighting_method = 1
)
}
\arguments{
\item{azimuth}{azimuth for sample variogram/covariance (nvariogram * 1 vector), degrees}

\item{dip}{dips for sample variogram/covariance (nvariog * 1 vector), degrees}

\item{nlag}{number of lags for each variogram/correlogram (nvariog * 1 vector)}

\item{tail}{tail variables (nvariog * 1 vector), an integer vector}

\item{head}{head variables (nvariog * 1 vector), an integer vector}

\item{gam}{dataframe of simple and cross variograms/covariances. The shape of the dataframe should be sum(nlag) rows x 3 columns}

\item{model}{the variogram/covariance model matrix with dimensions of the number of structures (n_structures) * 8}

\item{vartype}{the script handles traditional variograms (1) or centered covariances (2)}

\item{maxiterations}{maximum number of consecutive iterations with no accepted transitions}

\item{max_time_s}{maximum processing time in seconds}

\item{p_acc_0}{probability of accepting a non-favorable transition at iteration 0}

\item{p_acc_2k}{probability of accepting a non-favorable transition at iteration 2000}

\item{gen_vec_corr}{correlation between generated Gaussian vectors}

\item{eps_thresh}{absolute comparison threshold for machine accuracy)}

\item{weighting_method}{0:constant weight, 1: ~npairs, 2:~/distance 3: ~npairs/distance}
}
\value{
sills of nested structures (n_structures * n_fields^2 matrix) and weighted sum of squares for optimal fit
}
\description{
Fit Linear Model of Coregionalization
Ported from Emergy et al. 2009
}
