% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_vp_model}
\alias{fit_vp_model}
\title{Make a model of vp ~ confining + deviatoric stress for mapping}
\usage{
fit_vp_model(df)
}
\value{
a linear model of vp
}
\description{
Make a model of vp ~ confining + deviatoric stress for mapping
}
