% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_vs_model}
\alias{fit_vs_model}
\title{Make a model of vs ~ confining + deviatoric stress for mapping}
\usage{
fit_vs_model(df)
}
\arguments{
\item{df}{machine learning dataframe}
}
\value{
a linear model of vs
}
\description{
Make a model of vs ~ confining + deviatoric stress for mapping
}
