% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_model}
\alias{generate_model}
\title{Wrapper Function for model generation}
\usage{
generate_model(
  target,
  features,
  model_type,
  pset,
  prefix,
  seed = 1,
  train_ratio = 0.9,
  mbo_budget = 75L,
  resample_iters = 1000,
  importance_reps = 50
)
}
\arguments{
\item{target}{target for model (character)}

\item{features}{features for model (character vector)}

\item{model_type}{MLR embedded model type}

\item{pset}{MLR parameter set for model type}

\item{prefix}{named prefix of output directory}

\item{seed}{random seem for resampling}

\item{train_ratio}{ratio of train data to all data}

\item{mbo_budget}{budget for bayesian optimization}

\item{resample_iters}{number of iterations for resampling performance analysis}

\item{importance_reps}{number of iterations for permutation importance}
}
\value{
TRUE
}
\description{
Wrapper Function for model generation
}
