% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_models}
\alias{generate_models}
\title{Wrapper of the Wrapper Function for model generation}
\usage{
generate_models(
  mldata,
  feats,
  feat_prefix,
  targets = c("youngs_modulus", "poisson_ratio", "brittleness")
)
}
\arguments{
\item{mldata}{the data (dataframe)}

\item{feats}{features for model (character vector)}

\item{feat_prefix}{the prefix for the feature set}

\item{targets}{named prefix of output directory}
}
\value{
TRUE
}
\description{
Wrapper of the Wrapper Function for model generation
}
