% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{get_head_id}
\alias{get_head_id}
\title{Helper to map head id from gstat to LMC code, rowwise implemented}
\usage{
get_head_id(id, data_ids)
}
\arguments{
\item{id}{id value from gstat}
}
\value{
head
}
\description{
Helper to map head id from gstat to LMC code, rowwise implemented
}
