% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{get_las_headers}
\alias{get_las_headers}
\title{Get Las Headers and return cleaned column}
\usage{
get_las_headers(asciitxt)
}
\arguments{
\item{asciitxt}{las ascii object}
}
\description{
Retrieve the columns provided in the LAS header
}
