% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_resample_regr_res}
\alias{get_resample_regr_res}
\title{Calculate the bias and variance for a resample object on a regression}
\usage{
get_resample_regr_res(resample)
}
\arguments{
\item{resample}{resample object from mlr::resample, predict must be 'both'}
}
\value{
list of train/test bias and variance + resample object for rds stash
}
\description{
Calculate the bias and variance for a resample object on a regression
}
