% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{get_subset_log_data}
\alias{get_subset_log_data}
\title{Mappable function to read LAS files and subset to duvernay base and top}
\usage{
get_subset_log_data(filepath, study_locs, dv_base, dv_top)
}
\arguments{
\item{filepath}{full filepath}

\item{study_locs}{an sf dataframe with study locations and kelly bushing}

\item{dv_base}{a raster of Duvernay base}

\item{dv_top}{a raster of Duvernay top}
}
\description{
Mappable function to read LAS files and subset to duvernay base and top
}
