% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{get_tail_id}
\alias{get_tail_id}
\title{Helper to map tail id from gstat to LMC code, rowwise implemented}
\usage{
get_tail_id(id, data_ids)
}
\arguments{
\item{id}{id value from gstat}
}
\value{
tail
}
\description{
Helper to map tail id from gstat to LMC code, rowwise implemented
}
