% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{getCurveHeaders}
\alias{getCurveHeaders}
\title{getCurveHeaders}
\usage{
getCurveHeaders(asciitxt)
}
\arguments{
\item{asciitxt}{las ascii object}
}
\value{

}
\description{
Retrieve the columns provided in the LAS header
}
