% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{getLASList}
\alias{getLASList}
\title{getLASList}
\usage{
getLASList(folder_path)
}
\arguments{
\item{folder_path}{folderto look for las files}
}
\value{

}
\description{
get a list of las files from the folder path provided
}
