% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{getNumbersTable}
\alias{getNumbersTable}
\title{getNumbersTable}
\usage{
getNumbersTable(asciitxt, headers)
}
\arguments{
\item{asciitxt}{las ascii object}

\item{headers}{headers to be set}
}
\value{

}
\description{
sets / cuts the headers, and unnecessary characters from las ascii, returns numeric valued table
}
