% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{load_las_file}
\alias{load_las_file}
\title{Load LAS file into a table with a name ID and formatted headers}
\usage{
load_las_file(lasfile, short_uwi)
}
\arguments{
\item{lasfile}{path to an las file}

\item{short}{id of that file (user-defined)}
}
\description{
Load LAS file into a table with a name ID and formatted headers
}
