% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/geospatial.R
\name{make_3d_plot}
\alias{make_3d_plot}
\title{Make a 3D plot using Plot3D with a viridis colour scale and sf dataframe}
\usage{
make_3d_plot(sf_df, z_col, col_col, col_func = viridis(50), size = 0.1)

make_3d_plot(sf_df, z_col, col_col, col_func = viridis(50), size = 0.1)
}
\value{
plot

plot
}
\description{
Make a 3D plot using Plot3D with a viridis colour scale and sf dataframe

Make a 3D plot using Plot3D with a viridis colour scale and sf dataframe
}
