% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{make_curves_table}
\alias{make_curves_table}
\title{Make a numeric table by cutting the headers and unnecessary las characters}
\usage{
make_curves_table(asciitxt, headers)
}
\arguments{
\item{asciitxt}{las ascii object}

\item{headers}{headers to be set}
}
\description{
Make a numeric table by cutting the headers and unnecessary las characters
}
