% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{make_lmc_input}
\alias{make_lmc_input}
\title{Make an input list for the LMC algorithm from gstat}
\usage{
make_lmc_input(g, vmodel)
}
\arguments{
\item{g}{gstat object and data}

\item{vmodel}{common variogram model used to build gstat object}
}
\value{
lmc_input list (named for use in lmc function)
}
\description{
Make an input list for the LMC algorithm from gstat
}
