% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{make_lmc_model}
\alias{make_lmc_model}
\title{Make an LMC compliant model from the gstat common variogram model}
\usage{
make_lmc_model(vmodel)
}
\arguments{
\item{vmodel}{common variogram model used to build gstat object}
}
\value{
lmc compliant model
}
\description{
Make an LMC compliant model from the gstat common variogram model
}
