% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_pdp_plot}
\alias{make_pdp_plot}
\title{Make a partial dependence plot of all features}
\usage{
make_pdp_plot(i, predictor, feats)
}
\arguments{
\item{i}{index for mapping}

\item{predictor}{IML predictor}

\item{feats}{feature vector (char vec)}
}
\value{
individual pdp plot
}
\description{
Make a partial dependence plot of all features
}
