% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{model_type_to_cova}
\alias{model_type_to_cova}
\title{Convert gstat factors to LMC COVA model types, list applied}
\usage{
model_type_to_cova(mod_factor)
}
\arguments{
\item{mod_factor}{the factor from gstat for model type}
}
\value{
integer of COVA type
}
\description{
Convert gstat factors to LMC COVA model types, list applied
}
