% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_importance_interaction}
\alias{plot_importance_interaction}
\title{Plot importance and interation plot}
\usage{
plot_importance_interaction(predictor, prefix, repetitions = 500)
}
\arguments{
\item{predictor}{number of repetitions for permutation feature importance}

\item{prefix}{named prefix of output directory}
}
\value{
TRUE
}
\description{
Plot importance and interation plot
}
