% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_pdp}
\alias{plot_pdp}
\title{Plot partial dependence plot}
\usage{
plot_pdp(features, target, predictor, prefix)
}
\arguments{
\item{features}{feature vector (char vec)}

\item{target}{name of target (char)}

\item{predictor}{IML predictor}

\item{prefix}{named prefix of output directory}
}
\value{
TRUE
}
\description{
Plot partial dependence plot
}
