% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot residuals for train and test set}
\usage{
plot_residuals(train_predict, test_predict, prefix)
}
\arguments{
\item{train_predict}{output from predict(model, newdata=train)}

\item{test_predict}{output from predict(model, newdata=test)}

\item{prefix}{named prefix of output directory}
}
\value{
TRUE
}
\description{
Plot residuals for train and test set
}
