% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pr_correction_loss}
\alias{pr_correction_loss}
\title{Average of Mean Absolute Error and RMSE
of Wave Equation PR prediction with a correction factor
(to match statistical models)
Used for cross validated optimization}
\usage{
pr_correction_loss(correction_factor = 0.01, df)
}
\arguments{
\item{correction_factor}{scalar correction factor (numeric)}

\item{df}{the data (dataframe)}
}
\value{
mean absolute error
}
\description{
Average of Mean Absolute Error and RMSE
of Wave Equation PR prediction with a correction factor
(to match statistical models)
Used for cross validated optimization
}
