% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{pull_substring}
\alias{pull_substring}
\title{Check and pull a substring from LAS}
\usage{
pull_substring(asciitxt, sub.str)
}
\arguments{
\item{asciitxt}{las ascii object}

\item{sub.str}{pattern to search for}
}
\value{
character string
}
\description{
Check and pull a substring from LAS
}
