% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{read_file_to_char}
\alias{read_file_to_char}
\title{Read a file into an ASCII text file}
\usage{
read_file_to_char(file_name)
}
\arguments{
\item{file_name}{File path to LAS}
}
\value{
char string
}
\description{
Read a file into an ASCII text file
}
