% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{readFileIntoChar}
\alias{readFileIntoChar}
\title{readFileIntoChar}
\usage{
readFileIntoChar(file_name)
}
\arguments{
\item{file_name}{File path to LAS}
}
\value{

}
\description{
reads LAS into ascii string
}
