% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_files.R
\name{readLASFileIntoTable}
\alias{readLASFileIntoTable}
\title{readLASFileIntoTable}
\usage{
readLASFileIntoTable(lasfile, short)
}
\arguments{
\item{lasfile}{path to an las file}

\item{short}{id of that file (user-defined)}
}
\value{

}
\description{
Main function, reads lasfile provided, and adds a name ID as a 'short' variable (you can )
}
