% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error of a model}
\usage{
rmse(model)
}
\arguments{
\item{model}{model output with a residuals column / vector}
}
\value{
scalar RMSE
}
\description{
Root Mean Square Error of a model
}
