% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostats.R
\name{setrot}
\alias{setrot}
\title{Setup a rotated matrix of a geostatistical model
Ported from Emery at al. 2009}
\usage{
setrot(model, it, eps_thresh)
}
\arguments{
\item{model}{the geostatistical model matrix}

\item{it}{the specific rows (the nested model) to rotate}
}
\value{
rotated matrix of the model
}
\description{
Setup a rotated matrix of a geostatistical model
Ported from Emery at al. 2009
}
