% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wave_pr_w_correction}
\alias{wave_pr_w_correction}
\title{Wave Equation for Poisson Ratio with a correction factor}
\usage{
wave_pr_w_correction(df, correction_factor)
}
\arguments{
\item{df}{the data (dataframe)}

\item{correction_factor}{scalar correction factor (numeric)}
}
\value{
vector, predicted poisson ratio values
}
\description{
Wave Equation for Poisson Ratio with a correction factor
}
