% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ym_correction_loss}
\alias{ym_correction_loss}
\title{Average of Mean Absolute Error and RMSE
of Wave Equation YM prediction with a correction factor
(to match statistical models)
Used for cross validated optimization}
\usage{
ym_correction_loss(correction_factor = 1, df)
}
\arguments{
\item{correction_factor}{scalar correction factor (numeric)}

\item{df}{the data (dataframe)}
}
\value{
mean absolute error
}
\description{
Average of Mean Absolute Error and RMSE
of Wave Equation YM prediction with a correction factor
(to match statistical models)
Used for cross validated optimization
}
