% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{ggvariogram_map}
\alias{ggvariogram_map}
\title{Make a variogram map with ggplot}
\usage{
ggvariogram_map(formula, sp_df, cutoff = NA, width = NA, threshold = NA)
}
\value{
plot
}
\description{
Make a variogram map with ggplot
}
